+++ api/eventData/eventData.go
@@ -36,23 +36,23 @@ func MethodHandler(w http.ResponseWriter, r *http.Request) {
 	id := parts[5]
 
 	//Check if it is already in firebase
-	data, exist, err := db.DB.Get("Events", id)
-	if err != nil {
-		debug.ErrorMessage.Update(
-			http.StatusInternalServerError,
-			"TicketMaster.Handler() -> Database.get() -> Trying to get data",
-			err.Error(),
-			"Unknown",
-		)
-		debug.ErrorMessage.Print(w)
-		return
-	}
+	data, exist := db.DB.Get("Events", id)
 
 	if exist { //If in firebase, fetch data from firebase
 
 		//Storing locally
 		var info FirebaseStore
-		info.readData(data.Container)
+		err := info.readData(data["Container"].(interface{}))
+		if err != nil {
+			debug.ErrorMessage.Update(
+				http.StatusInternalServerError,
+				"TicketMaster.Handler() -> Converting String Date to time.Time format",
+				err.Error(),
+				"Unknown",
+			)
+			debug.ErrorMessage.Print(w)
+			return
+		}
 
 		//Convert the date string
 		layOut := "2006-01-02"
@@ -73,7 +73,7 @@ func MethodHandler(w http.ResponseWriter, r *http.Request) {
 
 		if overdue {
 			//delete the entry in the firebase
-			err = db.DB.Delete(id)
+			err = db.DB.DeleteEvent(id)
 			if err != nil {
 				debug.ErrorMessage.Update(
 					http.StatusInternalServerError,