+++ api/weatherHoliday/weatherHoliday.go
@@ -1,12 +1,13 @@
 package weatherHoliday
 
 import (
+	"bytes"
 	"encoding/json"
-	"main/api/countryData"
-	"main/api/geocoords"
 	"main/api/holidaysData"
+	"main/api/notification/weatherEvent"
 	"main/db"
 	"main/debug"
+	"main/dict"
 	"net/http"
 	"strings"
 )
@@ -17,8 +18,7 @@ type WeatherHoliday struct {
 	Location string `json:"location"`
 	URL string `json:"url"`
 	Frequency string `json:"frequency"`		// Every day or on date
-	Timeout int `json:"timeout"`			// Hours
-	ID string `json:"id"`
+	Timeout int64 `json:"timeout"`			// Hours
 }
 
 // Register a webhook
@@ -36,42 +36,9 @@ func (weatherHoliday *WeatherHoliday) Register(w http.ResponseWriter, r *http.Re
 		return
 	}
 
-	// Get the geocoords of the location
-	var locationCoords geocoords.LocationCoords
-	status, err := locationCoords.Handler(weatherHoliday.Location)
-	if err != nil {
-		debug.ErrorMessage.Update(
-			status,
-			"WeatherHoliday.Register() -> LocationCoords.Handler() -> Getting location info",
-			err.Error(),
-			"Unknown",
-		)
-		debug.ErrorMessage.Print(w)
-		return
-	}
-
-	// Get country and format it correctly
-	address := strings.Split(locationCoords.Address, ", ")
-	country := address[len(address)-1]
-
-	// Get country code
-	var countryInfo countryData.Information
-
-	status, err, countryCode := countryInfo.Handler(country)
-	if err != nil {
-		debug.ErrorMessage.Update(
-			status,
-			"WeatherHoliday.Register() -> CountryData.handler() -> Getting country code",
-			err.Error(),
-			"Selected country is not valid",
-		)
-		debug.ErrorMessage.Print(w)
-		return
-	}
-
 	// Get the country's holidays
 	var holidaysMap = make(map[string]interface{})
-	holidaysMap, status, err = holidaysData.Handler(countryCode)
+	holidaysMap, status, err := holidaysData.Handler(weatherHoliday.Location)
 	if err != nil {
 		debug.ErrorMessage.Update(
 			status,
@@ -84,30 +51,40 @@ func (weatherHoliday *WeatherHoliday) Register(w http.ResponseWriter, r *http.Re
 	}
 
 	// Make the first letter of each word uppercase
-	weatherHoliday.Holiday = strings.Title(weatherHoliday.Holiday)
+	weatherHoliday.Holiday = strings.Title(strings.ToLower(weatherHoliday.Holiday))
 
 	// Check if the holiday exists in the selected country
-	_, ok := holidaysMap[weatherHoliday.Holiday]
+	key, ok := holidaysMap[weatherHoliday.Holiday]
 	if !ok {
-		http.Error(w, "The selected holiday is not valid", http.StatusBadRequest)
+		debug.ErrorMessage.Update(
+			http.StatusBadRequest,
+			"WeatherHoliday.Register() -> Checking if a holiday exists in a country",
+			"invalid holiday: the holiday is not valid in the selected country",
+			"Not a real holiday. Check your spelling and make sure it is the english name.",
+		)
+		debug.ErrorMessage.Print(w)
 		return
 	}
 
-	// Check if the frequency field is valid
-	if weatherHoliday.Frequency != "ON_DATE" && weatherHoliday.Frequency != "EVERY_DAY" {
-		http.Error(w, "The selected frequency is not valid. Try writing either 'ON_DATE' or 'EVERY_DAY'", http.StatusBadRequest)
-		return
-	}
+	// Save data to weatherEvent struct
+	var data weatherEvent.WeatherEventInput
 
-	// Add webhook to the database
-	var dataDB db.Data
-	dataDB.Container = weatherHoliday
+	data.Date = key.(string)
+	data.Location = weatherHoliday.Location
+	data.URL = weatherHoliday.URL
+	data.Frequency = weatherHoliday.Frequency
+	data.Timeout = weatherHoliday.Timeout
 
-	_, id, err := db.DB.Add("WeatherHoliday", "", dataDB)
+	// Convert struct to json
+	jsonData, err := json.Marshal(data)
+
+	// Send data to weatherEvent endpoint
+	var weatherEvent weatherEvent.WeatherEvent
+	req, err := http.NewRequest("POST", dict.WEATHEREVENT_PATH, bytes.NewBuffer(jsonData))
 	if err != nil {
 		debug.ErrorMessage.Update(
 			http.StatusInternalServerError,
-			"WeatherHoliday.Register() -> db.Add() -> Adding webhook to the database",
+			"WeatherHoliday.Register() -> Sending request to weatherEvent",
 			err.Error(),
 			"Unknown",
 		)
@@ -115,16 +92,14 @@ func (weatherHoliday *WeatherHoliday) Register(w http.ResponseWriter, r *http.Re
 		return
 	}
 
-	weatherHoliday.ID = id
-
-	http.Error(w, "Webhook registered", http.StatusCreated)
+	weatherEvent.POST(w, req)
 }
 
 // Delete a webhook
 func (weatherHoliday *WeatherHoliday) Delete(w http.ResponseWriter, r *http.Request) {
 	// Parse URL path and ensure that the formatting is correct
 	path := strings.Split(r.URL.Path, "/")
-	if len(path) != 6 {
+	if len(path) != 7 {
 		debug.ErrorMessage.Update(
 			http.StatusBadRequest,
 			"WeatherHoliday.Delete() -> Parsing URL",
@@ -138,7 +113,7 @@ func (weatherHoliday *WeatherHoliday) Delete(w http.ResponseWriter, r *http.Requ
 	// Get webhook ID
 	id := path[len(path)-1]
 
-	err := db.DB.Delete(id)
+	err := db.DB.Delete("notifications", id)
 	if err != nil {
 		debug.ErrorMessage.Update(
 			http.StatusInternalServerError,