+++ dict/dictionary.go
@@ -4,8 +4,11 @@ package dict
 var URL = "localhost"
 var WEATHER_PATH = URL + "/weather-rest/v1/weather/location/"
 var WEATHERCOMPARE_PATH = URL + "/weather-rest/v1/weather/compare/"
+var WEATHEREVENT_PATH = URL + "/weather-rest/v1/event/date/"
 var EVENT_PATH = URL + "/weather-rest/v1/weather/event/"
 var DIAG_PATH = URL + "/weather-rest/v1/weather/diag/"
+var HOLIDAY_PATH = URL + "/weather-rest/v1/notification/weather/holiday/"
+var WEATHERHOOK_PATH = URL + "/weather-rest/v1/notification/weather"
 
 // GetWeatherURL generates Weather URL according to parameters
 func GetWeatherURL(location string, date string) string {