+++ ChangeLog.md
@@ -168,3 +168,20 @@
 
 [3.14.0]    **development**:    Added ticketmaster endpoint
 
+[3.15.0]    **development**:    Moved code from holiday webhook to holiday data
+
+[3.15.1]    **fix**:            Fixed an error when formatting holiday name wrong
+
+[3.16.0]    **development**:    Added WeatherEvent POST method and a feedback pipeline to handle webhook feedback
+
+[3.17.0]    **development**:    Added passing of data from weatherHoliday to weatherEvent
+
+[3.18.0]    **development**:    Added WeatherEvent GET method and modified database.Get() to return map of interfaces
+
+[3.18.1]    **fix**:            Fixed weatherHook to work with new database.Get()
+
+[3.18.2]    **development**:    Slight cleanup of db.Delete, more work on weatherHook
+
+[3.18.3]    **fix**:            Fixed WeatherHoliday to work with new database.Get()
+
+[3.19.0]    **fix**:            Added diag endpoint and restructured eventData